fix creating objects and stuff

+++ Assets/Scripts/Network/NetworkClient.cs
@@ -77,20 +77,20 @@ public class NetworkClient : MonoBehaviour, ISyncClient, INetworkMessageReceiver
     public void ReceiveNetworkMessage(NetworkMessage message)
     {
         SceneChangePacket sceneChange = NetworkUtils.GetMessage<SceneChangePacket>(message.contents);
-        if (sceneChange != null)
+        if (sceneChange != null && sceneChange.tag == typeof(SceneChangePacket).Name)
         {
             Debug.Log($"Scene change received: {sceneChange.scene}: {sceneChange.startingObjectID}.");
             SetScene(sceneChange.scene, sceneChange.startingObjectID);
             return;
         }
         CreateNetworkObjectPacket createObject = NetworkUtils.GetMessage<CreateNetworkObjectPacket>(message.contents);
-        if (createObject != null)
+        if (createObject != null && sceneChange.tag == typeof(CreateNetworkObjectPacket).Name)
         {
             networkObjects.CreateNetworkObject(createObject.objectID, createObject.networkType);
             return;
         }
         DestroyNetworkObjectPacket destroyObject = NetworkUtils.GetMessage<DestroyNetworkObjectPacket>(message.contents);
-        if (destroyObject != null)
+        if (destroyObject != null && sceneChange.tag == typeof(DestroyNetworkObjectPacket).Name)
         {
             networkObjects.DestroyNetworkObject(destroyObject.objectID);
             return;
